refactor(routines): complete RoutineDetail.tsx refactoring
- Split monolithic component into 7 smaller, reusable components
- Create RoutineDetailHeader, RoutineOverviewCards, RoutineProgress
- Add WeeklySchedule, RoutineInfo, WorkoutDaysList, ExerciseList
- Reduce file size from 547 to 200 lines (63% reduction)
- Improve maintainability and code organization
- Fix all linting and formatting issues
