+++ apps/web/src/features/routines/components/RoutineInfo.tsx
@@ -1,0 +1,86 @@
+'use client';
+
+import { Copy, Share, Trash } from 'lucide-react';
+
+import { Button } from '@/components/ui/button';
+import { Card } from '@/components/ui/card';
+import { getDifficultyColor } from '../utils';
+
+interface RoutineInfoProps {
+  difficulty: string;
+  createdDate: string;
+  lastModified: string;
+  onDuplicate: () => void;
+  onShare: () => void;
+  onDelete: () => void;
+}
+
+const RoutineInfo = ({
+  difficulty,
+  createdDate,
+  lastModified,
+  onDuplicate,
+  onShare,
+  onDelete,
+}: RoutineInfoProps): React.ReactElement => {
+  return (
+    <Card className="p-6">
+      <h2 className="text-lg font-semibold text-gray-800 mb-4">Routine Info</h2>
+      <div className="space-y-4">
+        <div>
+          <div className="text-sm text-gray-500 mb-1">Difficulty Level</div>
+          <span
+            className={`px-3 py-1 rounded-full text-sm font-medium ${getDifficultyColor(difficulty)}`}
+          >
+            {difficulty}
+          </span>
+        </div>
+
+        <div>
+          <div className="text-sm text-gray-500 mb-1">Created</div>
+          <div className="font-medium text-gray-800">{createdDate}</div>
+        </div>
+
+        <div>
+          <div className="text-sm text-gray-500 mb-1">Last Modified</div>
+          <div className="font-medium text-gray-800">{lastModified}</div>
+        </div>
+
+        <div className="pt-4 border-t border-gray-200">
+          <div className="text-sm text-gray-500 mb-3">Quick Actions</div>
+          <div className="space-y-2">
+            <Button
+              variant="outline"
+              size="sm"
+              className="w-full justify-start"
+              onClick={onDuplicate}
+            >
+              <Copy className="w-4 h-4 mr-2" />
+              Duplicate Routine
+            </Button>
+            <Button
+              variant="outline"
+              size="sm"
+              className="w-full justify-start"
+              onClick={onShare}
+            >
+              <Share className="w-4 h-4 mr-2" />
+              Share Routine
+            </Button>
+            <Button
+              variant="outline"
+              size="sm"
+              className="w-full justify-start text-red-600 hover:text-red-700"
+              onClick={onDelete}
+            >
+              <Trash className="w-4 h-4 mr-2" />
+              Delete Routine
+            </Button>
+          </div>
+        </div>
+      </div>
+    </Card>
+  );
+};
+
+export default RoutineInfo;
