+++ apps/web/src/features/routines/index.ts
@@ -10,6 +10,15 @@ export { default as WorkoutCard } from './components/WorkoutCard';
 export { default as RoutinesList } from './components/RoutinesList';
 export { default as ActiveRoutineCard } from './components/ActiveRoutineCard';
 
+// Routine detail components
+export { default as RoutineDetailHeader } from './components/RoutineDetailHeader';
+export { default as RoutineOverviewCards } from './components/RoutineOverviewCards';
+export { default as RoutineProgress } from './components/RoutineProgress';
+export { default as WeeklySchedule } from './components/WeeklySchedule';
+export { default as RoutineInfo } from './components/RoutineInfo';
+export { default as WorkoutDaysList } from './components/WorkoutDaysList';
+export { default as ExerciseList } from './components/ExerciseList';
+
 // Workout creation components
 export { default as WorkoutHeader } from './components/WorkoutHeader';
 export { default as WorkoutDetails } from './components/WorkoutDetails';
